Fix length rounding
... also add const in one place

Amends: 091cc85d4aec9bc147923975f42cc71307d6fe32

+++ src/gui/PlaylistWidget.hpp
@@ -51,7 +51,7 @@ private:
 		QImage img;
 
 		bool updateLength;
-		int length;
+		double length;
 	};
 private slots:
 	void updateItem(ItemUpdated iu);
